fix(frontend): fix type annotations, move interfaces to `book.model.ts`, and add metadata-related service methods
