+++ booklore-ui/src/app/features/book/components/book-browser/metadata-restore-dialog-component/metadata-restore-dialog-component.ts
@@ -30,10 +30,10 @@ export class MetadataRestoreDialogComponent implements OnInit {
     this.book = this.bookService.getBookByIdFromState(this.bookId);
 
     this.bookService.getBackupMetadata(this.bookId).subscribe({
-      next: (data) => {
+      next: (data: BookMetadata) => {
         this.backupMetadata = data;
       },
-      error: err => {
+      error: (err: any) => {
         this.messageService.add({
           severity: 'warn',
           summary: 'No Backup Found',
@@ -54,7 +54,7 @@ export class MetadataRestoreDialogComponent implements OnInit {
         });
         this.dynamicDialogRef.close({ action: 'restore', bookId: this.bookId });
       },
-      error: err => {
+      error: (err: any) => {
         const errorMessage = err?.error?.message || err?.message || 'Unknown error';
         this.messageService.add({
           severity: 'error',
